+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/api_args.h
@@ -3377,6 +3377,14 @@ typedef union rocprofiler_hip_api_args_t
         size_t*     paramSize;
     } hipKernelGetParamInfo;
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 21
+    struct
+    {
+        hipMemPool_t* memPool;
+        hipMemLocation* location;
+        hipMemAllocationType type;
+    } hipMemGetDefaultMemPool;
+#endif
 } rocprofiler_hip_api_args_t;
 
 ROCPROFILER_EXTERN_C_FINI